+++ doc/changes.md
@@ -1,5 +1,6 @@
 # Change Log # {#changes}
 
+  - Revised the matroid decomposition structure.
   - Fixed a bug in the code for printing dot-files for transposed graphic/network matrices; reported by Christopher Hojny.
   - Added a `timeLimit` parameter to all potentially time intensive functions.
   - Bugfix for the `timeLimit` parameter.